refactor: dto 패키지 controller 내부로 이동
