+++ src/main/java/roomescape/reservation/controller/dto/CreateReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.dto;
+package roomescape.reservation.controller.dto;
 
 import java.time.LocalDate;
 import roomescape.reservation.domain.Reservation;
