+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,10 +3,10 @@ package roomescape.reservation.service;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.stereotype.Service;
+import roomescape.reservation.controller.dto.CreateReservationRequest;
+import roomescape.reservation.controller.dto.ReservationResponse;
 import roomescape.reservation.dao.ReservationDao;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.dto.CreateReservationRequest;
-import roomescape.reservation.dto.ReservationResponse;
 import roomescape.reservationtime.dao.ReservationTimeDao;
 import roomescape.reservationtime.domain.ReservationTime;
 import roomescape.reservationtime.exception.ReservationTimeException;
